Add support for a pluggable logger.
This patch adds a generic call to get a logger, and a default
implementation, which uses the standard python logging libraries.  The
user can specify to the getLogger(config) method call which logging
system they want by providing a 'driver' entry in their config file.
This should be the fully qualified module name, and can be anywhere that
python can see it.  For the default driver, you can set any settings
supported by the dictConfig method, which is to say, all python logging
configuration.

Signed-off-by: Jason Guiditta <jason.guiditta@gmail.com>