Adding a dlrn sub-module
This is the start of a dlrn module the will start to drive atkinson.
This portion handles a few of the basic http requests that we have to
make to extract version data and what the current commit is for a given
promoted link.

We now rely on toolchest version 0.0.4 or greater for some of the
functionality.

Signed-off-by: Jason Joyce <fuzzball81@gmail.com>