Fixing an issue where dlrn extended hashes were not supported.
When an extended hash is in use, DlrnHttpData did not generate the
proper URL to fetch data.

Refactored the commit and get_versions methods into properties.

Added a link property to the dlrn config files to set the initial dlrn symlink
to use. This can been overridden in the constructors.

Signed-off-by: Jason Joyce <fuzzball81@gmail.com>