Adding a dlrn failed to build from source error.
The first error that you can encounter from dlrn is packages failing to
build. The DlrnFTBFSError is the handler for these types of errors. The
class files and maintains the error through it's lifecycle. This type of
error needs a package maintainer to look into it, so the action is
simply to report the error via a reporter.

A BaseError pattern is also established so future errors can be handled
in a consistent manner.

Signed-off-by: Jason Joyce <fuzzball81@gmail.com>