Adding a trello based error reporter.
The trello reporter creates and maintains trello cards via the trello
rest api. When a new report is created a card is added to the board
based on the configuration file. The report also has support for adding
checklist items to track progress.

The configuration file needs the api key and token from trello to be
able to access and update cards.

Signed-off-by: Jason Joyce <fuzzball81@gmail.com>