Only run a browser on systems with a GUI. (#20)
If a developer does not have a GUI running then
decline running the browser but perform the rest of the actions in the
target. If a command called a browser on a headless system an error could
occur. This checks for a display and if one is found in the environment
then launch the browser.

Signed-off-by: Jason Joyce <fuzzball81@gmail.com>