Using the dlrn status report to detect failed builds.
To get failure information, parsing the status_report.csv file is
the easiest way to get a list of builds that are failing. The csv file
provides all of the data needed to generate information to report the
error.

The unit tests use both good and bad csv files generated from a dlrn
instance.

Signed-off-by: Jason Joyce <fuzzball81@gmail.com>